Splitted main exucution code into 3 functions, split_files_from_path, start_mapper_threads, launch_reducer_threads. This will be easier to implement threads as we will be able to call other function if we want to perfom map and/or reduce task on sockets and/or threading
